+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes relevant to players in this project will be documented in th
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/). This project uses its own versioning system.
 
 ## [Unreleased]
+### Changed
+- Heroes
+  - Scholar now requires 2 unlocked techs to unlock (from 7,500 research points)
+  - Scion now requires 6 successful attacks to unlock (from 500 prestige)
 
 ## [1.47.0] - 2025-10-31
 ### Changed